# backEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piano Shop Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/brand_create": {
            "post": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for insert brand to database",
                "operationId": "7bc0bab47e75015ab185188576dffe6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter Brand Name",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Enter Brand Image Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/brand_update/{id}": {
            "put": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for update brand information",
                "operationId": "a81f23f65b4aa22faf97435fa9cc7441",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter brand Name",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Enter brand Image Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Brand updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Brand not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/brand_delete/{id}": {
            "delete": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for delete brand from database",
                "operationId": "22f3646d5c5b05edd3c5ea13eb99604e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/brand_show_by_id/{id}": {
            "get": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for get brand by id",
                "operationId": "b8bd6d1d4070924630a35a6717733257",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Brand not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/category_create": {
            "post": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for insert category to database",
                "operationId": "640c1192441a3a3624f0115b54f85a79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter Brand Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/category_update/{id}": {
            "put": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for update category information",
                "operationId": "dad1b90865e323966f5cf22cf8e96ef6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter category Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/category_delete/{id}": {
            "delete": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for delete category from database",
                "operationId": "224f99835467e9fb53323c131f4219b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/category_show_by_id/{id}": {
            "get": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for get category by id",
                "operationId": "5c143e1d9d5303e9b7cd33e68a595254",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/color_create": {
            "post": {
                "tags": [
                    "Product Details Api"
                ],
                "description": "use for insert color to database",
                "operationId": "b62632d475810b478e9ef1bb4ff7f36b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter Color Name",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Enter Color Image Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}